backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
openpyxl
msal
python-jose[cryptography]
passlib[bcrypt]
python-multipart
